* io.c (io_unread): cast as long the value for extra_max.
	  [ruby-core:44137] [Bug #6257]


git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/branches/ruby_1_9_3@35256 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ io.c
@@ -290,7 +290,7 @@ inline static int set_binary_mode_with_seek_cur(rb_io_t *fptr) {
 	return setmode(fptr->fd, O_BINARY);
     }
     /* add extra offset for removed '\r' in rbuf */
-    extra_max = pos - fptr->rbuf.len;
+    extra_max = (long)(pos - fptr->rbuf.len);
     p = fptr->rbuf.ptr + fptr->rbuf.off;
     for (i = 0; i < fptr->rbuf.len; i++) {
 	if (*p == '\n') newlines++;
